archive/SpreadWork/requirements.txt
google-cloud-datastore==1.10.0
google-cloud-pubsub==1.0.2
